+++ scratch.js
@@ -1,6 +1,4 @@
 const spacetime = require('./src/index')
 
-console.time('time')
-let s = spacetime('june 20 2019')
-s.week()
-console.timeEnd('time')
+let s = spacetime('june 20 2019', '-5:00')
+console.log(s.timezone())